% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_DB_helpers.R
\name{append_table_sql_lite}
\alias{append_table_sql_lite}
\title{This function will upload new rows to an existing table in a sql lite
DB file. The dataframe object provided must have the same column
names as the column names in the DB.}
\usage{
append_table_sql_lite(.data, table_name, conn)
}
\arguments{
\item{.data}{(dataframe) The data you would like to append to the
SQL lite database.}

\item{table_name}{(character) The name you would like to give the table
you are creating in the database.}

\item{conn}{}
}
\value{
(NULL)
}
\description{
This function will upload new rows to an existing table in a sql lite
DB file. The dataframe object provided must have the same column
names as the column names in the DB.
}
\examples{
\dontrun{
#' local_db_path <- "data/joined_lm_trig_data/mydb.db"
db_con <- connect_db_local(local_db_path)

table_name <- "dummy"
data_to_upload <- 
  tibble(
  test1 = c(1,2,3), 
  test2= c("a", "b", "c")
  )
  
append_table_sql_lite(
  .data = data_to_upload, 
  table_name = table_name, 
  conn = db_con
)   

}
}
