% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AusTender.R
\name{aus_tender_yearly_spend}
\alias{aus_tender_yearly_spend}
\title{Convert spend to spend per year by start and end date for aus tender data.}
\usage{
aus_tender_yearly_spend(.data = get_austender_local_raw())
}
\arguments{
\item{.data}{(tibble) Needs x07_end_date, x06_start_date and x39_value}
}
\value{
(tibble) Returns the input Aus Tender data with a cost per year
column.
}
\description{
Convert spend to spend per year by start and end date for aus tender data.
}
\examples{
\dontrun{

raw_tender <- get_austender_local_raw()

raw_data_desc_cleaned <- 
  clean_description(
  .data =raw_tender
  descrip_col = "x10_description", 
  new_col_name "descrip_removed"
  )
  
raw_data_with_yearly_spend <- 
     aus_tender_yearly_spend(
     .data = raw_data_desc_cleaned
     )

}
}
