% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AusTender.R
\name{clean_description_aus_tender}
\alias{clean_description_aus_tender}
\title{Cleans the description column, removes uneccessary characters}
\usage{
clean_description_aus_tender(
  .data = get_austender_local_raw(),
  descrip_col = "x10_description",
  new_col_name = "descrip_removed"
)
}
\arguments{
\item{.data}{(tibble) Use get_austender_local_raw() as the input.}

\item{descrip_col}{description column to clean}

\item{new_col_name}{new description column name (default: "descrip_removed")}
}
\value{
(tibble) Returns the input data set with the descriptions cleaned.
}
\description{
Cleans the description column, removes uneccessary characters
}
\examples{
\dontrun{

raw_tender <- get_austender_local_raw()
raw_data_desc_cleaned <- 
  clean_description(
  .data =raw_tender
  descrip_col = "x10_description", 
  new_col_name "descrip_removed"
  )

}
}
