% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_DB_helpers.R
\name{connect_db_local}
\alias{connect_db_local}
\title{This function will create a local/connect to a local sql lite file. The
path specified will be where the sql lite database is created to the
local database that the function will connect you to.}
\usage{
connect_db_local(path)
}
\arguments{
\item{path}{(file path character) File path for where you want the sql lite
database file to be created or the path containing the sql lite database
you would like to connect to.}
}
\value{
(Database connection object)
}
\description{
This function will create a local/connect to a local sql lite file. The
path specified will be where the sql lite database is created to the
local database that the function will connect you to.
}
\examples{
\dontrun{
local_db_path <- "data/joined_lm_trig_data/mydb.db"
db_con <- connect_db_local(local_db_path)

}
}
