% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_drive_helper.R
\name{create_one_drive_path}
\alias{create_one_drive_path}
\title{Create One Drive file Path}
\usage{
create_one_drive_path(
  user = "Nikhil Chandra",
  team = "OneDrive - 35 South",
  path_extension = "raw data/abs_geo_structures"
)
}
\arguments{
\item{user}{(character) This is your user name. Example: "Nikhil Chandra"}

\item{team}{(character) This is the Team name on sharepoint. The default is
"OneDrive - 35 South", it is unlikley you will have to change this.}

\item{path_extension}{(character) This should be the path after
"C:/Users/Nikhil Chandra/OneDrive - 35 South/Documents - Data Science/". This
is the base path, if your file is in the folder raw_data/abs_anzic_income_SA2
with the file name "sa2_income_industry.csv" then you would provide
"raw_data/abs_anzic_income_SA2/sa2_income_industry.csv"}
}
\value{
(file path) This function will return the file path based on your
inputs.
}
\description{
This function will return a file path for your one drive path. You must provide
it your user name for it to create the appropriate path. You will also need
to provide the path extension which the part of the path after the root
'Documents - Data Science' directory in
}
\examples{
\dontrun{

file_path <- 
  create_one_drive_path(
  user = "Nikhil Chandra",
  team = "OneDrive - 35 South",
  path_extension = "raw_data/abs_anzic_income_SA2/sa2_income_industry.csv"
  )
  
data_read<- read_csv(file_path)

}
}
