% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AusTender.R
\name{detect_ict_descrip}
\alias{detect_ict_descrip}
\title{This function detects ICT words to detect various ICT spend categories inside
Aus Tender data. Use get_austender_local_raw() to get raw data and clean the
description column in the data using clean_description_aus_tender().}
\usage{
detect_ict_descrip(.data, descrip_col = "descrip_removed")
}
\arguments{
\item{.data}{(tibble) The data with the column that you want to have
categorized. Must have x13_category_title in the columns of the tibble.}

\item{descrip_col}{(character; "column") The name of the column to transform}
}
\value{
(tibble) Returns a dataframe with the added columns.
}
\description{
The function will categorise ICT spending in the main data set.
}
\examples{
\dontrun{

raw_tender <- get_austender_local_raw()

raw_data_desc_cleaned <- 
  clean_description(
  .data =raw_tender
  descrip_col = "x10_description", 
  new_col_name "descrip_removed"
  )
  
raw_data_with_yearly_spend <- 
     aus_tender_yearly_spend(
     .data = raw_data_desc_cleaned
     )
     
ict_categorised <- 
      detect_ict_descrip(.data = raw_data_with_yearly_spend, 
                          descrip_col = "descrip_removed")

}
}
