% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_drive_helper.R
\name{detect_user_onedrive}
\alias{detect_user_onedrive}
\title{Detect the Users One Drive}
\usage{
detect_user_onedrive()
}
\value{
(named character vector)
}
\description{
This function will detect your one drive path.
}
\examples{
\dontrun{

one_drive_path <- detect_user_onedrive()

}
}
