% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_helpers.R
\name{geo_code_aus_data_sa2}
\alias{geo_code_aus_data_sa2}
\title{This function will take a dataframe with coordinate data (-lat,long) and
intersect it with ABS maps data and return a dataframe with SA2 joined on.}
\usage{
geo_code_aus_data_sa2(
  .data,
  coordinate_col = "Coordinates",
  abs_data = sa2_geos
)
}
\arguments{
\item{.data}{(dataframe) A dataframe with coordinate data you want intersected
with abs maps data.}

\item{coordinate_col}{(character) The name of the column that contains
coordinate data.}

\item{abs_data}{(sf) An sf class data frame with a geometry column.}
}
\value{
(tibble)
}
\description{
This function will take a dataframe with coordinate data (-lat,long) and
intersect it with ABS maps data and return a dataframe with SA2 joined on.
}
\examples{
\dontrun{

cfps_data <- 
  create_one_drive_path(user = "Nikhil Chandra",
  path_extension = "raw data/coal_fire_power_stations/cfps_manual_web_scrape.csv") \%>\%
  read_csv()
sa2_geos <- absmapsdata::sa22021

cfps_with_sa2_2021<-
geo_code_aus_data_sa2(.data = cfps_data,
                      coordinate_col = "Coordinates",
                      abs_data = sa2_geos2)

}
}
