% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AusTender.R
\name{get_austender_cleaned}
\alias{get_austender_cleaned}
\title{Gets cleaned AusTender stored on One Drive}
\usage{
get_austender_cleaned(
  base_path = create_one_drive_path(path_extension =
    "raw data/AusTender_data/joined_data/aus_tender.csv")
)
}
\arguments{
\item{base_path}{(character) Path where the data is kept.}
}
\value{
(tibble) Returns a tibble of the read in data.
}
\description{
Gets cleaned AusTender stored on One Drive
}
\examples{
\dontrun{
aus_tender_data <- get_austender_cleaned()
}
}
