% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AusTender.R
\name{get_austender_local_raw}
\alias{get_austender_local_raw}
\title{Get AusTender Data}
\usage{
get_austender_local_raw(
  base_path = create_one_drive_path(path_extension =
    "raw data/AusTender_data/raw_exports")
)
}
\arguments{
\item{base_path}{(file path, character) This is the file path where the
data is kept. By default it uses
create_one_drive_path(path_extension = "raw data/AusTender_data/raw_exports").
But you may need to specify the file path if the function doest work with
its default variables.}
}
\value{
(tibble) This function will return the AusTender data into your
global environment
}
\description{
This function read in AusTender Data stored on share hub/One Drive. The data
is read in from csv files kep on sharehun
}
\examples{
\dontrun{

base_path <- 
    create_one_drive_path(
          user = NULL,
          path_extension = "raw data/AusTender_data/raw_exports"
          )
          
aus_tender_data <- 
    get_austender_local(
      base_path = base_path
    )
    
#Other Options

base_path <- 
    create_one_drive_path(
          user = janet,
          path_extension = "raw data/AusTender_data/raw_exports"
          )
          
aus_tender_data <- 
    get_austender_local(
      base_path = base_path
    )    

}
}
