% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpeR-package.R,
%   R/helperfunctions35South-package.R
\docType{package}
\name{helperfunctions35South-package}
\alias{helperfunctions35South}
\alias{helperfunctions35South-package}
\title{helperfunctions35South: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
