% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_DB_helpers.R
\name{write_table_sql_lite}
\alias{write_table_sql_lite}
\title{This function will take a dataframe and upload that dataframe as a table
to a SQL lite database object. You specify the name of the table you would
like to create and it will create that table in the SQL lite DB file and
upload the dataframe as the data. The column names of the new table will
be specified by the column names in the dataframe uploaded.}
\usage{
write_table_sql_lite(.data, table_name, conn)
}
\arguments{
\item{.data}{(dataframe) The data you would like to upload to the
SQL lite database.}

\item{table_name}{(character) The name you would like to give the table
you are creating in the database.}

\item{conn}{(Database connection object) Database connection see
\code{connect_db_local}}
}
\description{
This function will take a dataframe and upload that dataframe as a table
to a SQL lite database object. You specify the name of the table you would
like to create and it will create that table in the SQL lite DB file and
upload the dataframe as the data. The column names of the new table will
be specified by the column names in the dataframe uploaded.
}
\examples{
\dontrun{
#' local_db_path <- "data/joined_lm_trig_data/mydb.db"
db_con <- connect_db_local(local_db_path)

table_name <- "dummy"
data_to_upload <- 
  tibble(
  test1 = c(1,2,3), 
  test2= c("a", "b", "c")
  )
  
write_table_sql_lite(
  .data = data_to_upload, 
  table_name = table_name, 
  conn = db_con
)   

}
}
